Stop publishing some artifacts
We used to publish a tarball that contained both the provider and the
node library because we used them when building in other repositories,
back when we still used `yarn link` everywhere.

We no longer do this and can stop publishing these artifacts. We
continue to publish the provider plugin and NPM/PyPI plugins as we
used to.