Adopt breaking change from AWS provider (#84)
The `port` property can now be undefined on aws TargetGroups when a Lambda is targetted.  It is not possible for the TargetGroup created by this wrapper to create the Lambda type yet though, so we can just assert the undefined away.

Also updates dependencies to fixed released versions for package and floating released versions for tests - instead of `dev`.