Make aws.ec2.Vpc instance a child of awsx.ec2.Vpc
With `parent` unset, the aws.ec2.Vpc resource which is created by `awsx.ec2.Vpc`
components is created as a top-level resource with version 0.17.13 of the Pulumi
CLI and the latest version of the package. This leaves the tree rendered as part
of `pulumi {preview,update}` somewhat confusing if you have several instances of
the class, since not all resources are nested under the component which created
them.