README change to rename security group to valid name
Renames the security group to a valid name and adds the `awsx.ec2` namespace.

As "sg" the security group failed to be created with an error from AWS:
```
error: Error creating Security Group: InvalidParameterValue: Value (sg-017c251) for parameter GroupName is invalid. Group names may not be in the format sg-*.
```