Update ecs types
The awsx.ecs types have drifted a little relative to the aws.ecs types.  This change brings them up to date with the latest types in the AWS package.  In particular this adds:
* Container#interactive
* ServiceArgs#forceNewDeployment
* TaskDefinitionArgs#inferenceAccelerators
* TaskDefinitionArgs#ipcMode
* TaskDefinitionArgs#pidMode
* TaskDefinitionArgs#proxyConfiguration

Aside: The implementation includes an Override type that attempts to derive the correct type from the upstream types.  If this had been used directly, it would have avoided the need for this maintenance.  But documentation and error message experiences in TypeScript are not good for these derived types, so we instead also manually maintain a copy of the types.  We then attempt to verify that the manually maintained copy is compatible, but that attempt only verifies required properties, leaving the possibility that drift in optional properties (which in practice is the only kind of drift that can happen) will not be caught.  We likely need to revisit the overall strategy here.

Fixes #615.