Allow constructing Vpc with no args object
Since all input properties are optional, make the `args` bag itself optional, to align with the code-generation for custom resources with all-optional args bags.

Because the test infrastructure in this repo currently requires passing explict providers, which means passing an options bag, there is unfortunately no simple way to test this directly.