Fix continueBeforeSteadyState
The docs say this should default to `false`, and the implementation appears to be trying to implement that, but it applies `!` to an `Output` value, which will always be `false`, even if the `Output` internally evaluates to `false`.  Instead, apply the `!` inside an apply.

Also remove accidentally added debug logging.

Fixes #902.