Remove explicit dotnet install
We're installing 3.1 on an image which already has 6.0. This has no effect as dotnet will automatically choose the newest version.

This was found because were getting an error on the acceptance tests where we used the install action but failed to specify the matrix.dotnetversion which meant the action automatically installed the latest version (7.0) which caused build errors.