Start adding upgrade tests (#1110)
This PR introduces a new kind of coverage, tests that check if upgrading from a baseline to the current version of the provider is safe for the user (no unexpected replacements). Specifically the change include:

- new ./provider Go module to host these style of tests
- dependency on pulumi/providertest
- `make test_provider` now runs a language-sdk-agnostic (yaml) subset of new tests
- `make test_python`, `make test_nodejs` and so on now run language-specific subsets of the new tests
- CI is modified to separate building from testing targets; AWS credentials are added to the building phase to support preview-only mode of upgrade tests that may require credentials (although they still run fast and do not provision infra)
- gotestfmt is installed in the Makefile instead of in GitHub actions to workaround a clobbering issue
- three select examples are fixed to be compatible with `cd ./examples/xyz && pulumi preview` by adding missing dependencies

This PR is pertinent to https://github.com/pulumi/pulumi-awsx/issues/1108 but does not completely fix it yet as not enough critical resources are covered.