Fix hard coded subnet masks
Using hard coded /19, /20 and /24 masks doesn't work when the VPC range is too small. We result in negative "new bits" and therefore subnets which are bigger than the available space.
We need to maintain backward compatibility, so anyone generating valid subnets right now will continue with the same results.
Therefore, we use an alternative, calculated mask if the minimal size block (largest mask number) is larger than the hard-coded value.
Add a test to validate all combinations (144 possibilities) result in subnets which are not larger than the VPC's original CIDR block. This is a pretty crude test, but was used to find the deficient code.