Correctly support apiKeySource in classic provider (#1159)
We were previously passing apiKeySource only via Swagger.  Even though AWS API Gateway does include the API key source in the swagger on output, it apparently does not read it from the swagger on updates.

However, it does support passing it outside of the Swagger, so we just pass this along directly to the RestAPI constructor is specified.

We could in principle *stop* sending it in the Swagger as well (since I don't think it's read from there at all), but I am avoiding changing that in case it has other subtle impacts on behaviour.