fix: targetGroup errors appear in containerDefinition diff
The [container
definition](https://www.pulumi.com/registry/packages/aws/api-docs/ecs/taskdefinition/#containerdefinitions_nodejs)
property of a Task Definition is a JSON string. We do not provide types
for the container definition which means it is possible to include
things in that JSON string that shouldn't be there. In the case of
this bug, we are including an additional `targetGroup` object in the
container definition string.

To fix this, I am removing the `targetGroup` field at the time that we
render the container definition JSON string. Ideally we may want to
created an interface type for a container definition in the AWS provider
which would make these errors easier to catch.

fix #1248