Modify weekly pulumi upgrade action to run acceptance tests in PR (#1313)
Previously the weekly pulumi upgrade couldn't auto-merge because it used
the built-in action access token for creating the PR. That default GH
action user doesn't have the necessary access rights (i.e. is not a
member of pulumi) to run acceptance tests.

This change makes the gh cli use the pulumi bot token.