Use v3 of actions/download-artifact (#1325)
https://github.com/pulumi/pulumi-awsx/pull/1323 used v4 of the download
artifact action, which is incompatible with uploads using v3 of the
upload action. See
[FAQ](https://github.com/actions/toolkit/blob/main/packages/artifact/docs/faq.md#which-versions-of-the-artifacts-packages-are-compatible)
for compatibility matrix.

This resulted in the release step
[failing](https://github.com/pulumi/pulumi-awsx/actions/runs/9620941727/job/26540766895#step:4:36)
to download the uploaded artifact.

This PR also updates the pre-release workflow while we're at it.

Note: I've opted to use v3 of the download action to maintain
consistency with the rest of our GHA workflows. We can update to v4 in
the future.