Rewrite a warning to avoid printing sub-properties as outputs (#1395)
Another fix for https://github.com/pulumi/pulumi-awsx/issues/1372 while
https://github.com/pulumi/pulumi-awsx/pull/1374 only partially
succeeded.

The previous attempt was successful in resolving the top-level output
but child outputs were still unresolved. I was a bit surprised that was
the case... outputs are hard!

The new attempt relies on `pulumi.jsonStringify` which is made exactly
for this purpose. This drops JSON formatting but hopefully that not too
bad for a small JSON. Example of the result:

```
 warning: The default subnetStrategy will change from "Legacy" to "Auto" in the next major version. Please specify the subnetStrategy explicitly. The current subnet layout can be specified via "Auto" as:
    
    [{"cidrMask":20,"type":"Public"}]
```