chore: fix renovate (#1465)
Hopefully this one works. Copied more of what we are doing with
pulumi-eks. Couple of things this does

1. Installs the v1 yarn (renovate container has v4 installed)
2. Installs pulumictl (copied from pulumi-eks)

I ran this inside the renovate docker container and it didn't fail, not
sure if there is a better way of testing this.

```
docker pull renovate/renovate:39-full
docker run -it -v $PWD:/usr/src/app renovate/renovate:39-full make renovate
```

re #1463