chore: ensure deduplicated dependencies (#1482)
This is an important check to include to fail early. If Renovate updates
dependencies in a way that causes several copies of `@pulumi/pulumi` SDK
to be included then the provider fails at runtime with an obscure error
about invokes, as in https://github.com/pulumi/pulumi-awsx/pull/1479 -
this will fail the build early instead.