[vpc/vpcEndpoints] use shared tags & expose endpoints (#1580)
👋 I had some internal tagging policies fail when using the awsx.ec2.Vpc.
I included the shared tags for the failing children (vpc endpoints)
here. When testing it, I discovered that the vpc weren't being added to
the component data, so I added that too. Along the way, I wrote some
tests for other things that might have had the same problem, and I left
them because they seem benign enough, and maybe even helpful.

I'm not familiar with this codebase, but it looks reasonable that
`build_sdks` produced no output.