sort TaskDefinitionContainerDefinitionInputs by container name to reduce non-needed deploys (#1626)
My team has been running into issues with unnecessary container task
definition updates for our services which causes lost time in waiting
for these changes to be deployed.

The cause of this was having multiple containers where the containers
would be getting mapped to the task definition in a non-predictable
manor which lead to false positive changes.

This PR addresses that issue by ensuring we sort the containers for a
task definition by their name in an effort to reduce these false
positive changes from occurring.