Fix cloudtrail to cloudwatch logs access
It doesn't look like the CloudTrail => CloudWatch Logs integration ever
worked for awsx. This PR fixes that by created an IAM role with the
necessary permissions.

Also, previously we had the property name incorrect, which means it
didn't even try to setup the log group.

Added a new test to assert that we create everything correctly, and
updated an existing test to show that we don't create the role if we
don't create the group.

fixes #1745